Move ProfileTab.title to l10n
