+++ lib/routes.dart
@@ -130,28 +130,7 @@ enum ProfileTab {
   sections,
   download,
   danger,
-  logout;
-
-  /// Returns localized title of this [ProfileTab].
-  String title() {
-    return switch (this) {
-      ProfileTab.public => 'label_profile'.l10n,
-      ProfileTab.signing => 'label_login_options'.l10n,
-      ProfileTab.link => 'label_link_to_chat'.l10n,
-      ProfileTab.background => 'label_background'.l10n,
-      ProfileTab.chats => 'label_chats'.l10n,
-      ProfileTab.calls => 'label_calls'.l10n,
-      ProfileTab.media => 'label_media'.l10n,
-      ProfileTab.notifications => 'label_notifications'.l10n,
-      ProfileTab.storage => 'label_storage'.l10n,
-      ProfileTab.language => 'label_language'.l10n,
-      ProfileTab.blocklist => 'label_blocked_users'.l10n,
-      ProfileTab.sections => 'label_show_sections'.l10n,
-      ProfileTab.download => 'label_download'.l10n,
-      ProfileTab.danger => 'label_danger_zone'.l10n,
-      ProfileTab.logout => 'btn_logout'.l10n,
-    };
-  }
+  logout,
 }
 
 /// Application's router state.
